Moved Picker props to its own module (#2065)
