+++ packages/components/src/spectrum/picker/index.ts
@@ -1,4 +1,6 @@
 export * from './Picker';
 export * from './PickerNormalized';
+export * from './PickerProps';
 export * from './usePickerItemScale';
+export * from './usePickerProps';
 export * from './usePickerScrollOnOpen';
