+++ packages/components/src/spectrum/utils/itemUtils.ts
@@ -110,55 +110,6 @@ export type NormalizedItemOrSection<TItemOrSection extends ItemOrSection> =
 
 export type NormalizedSpectrumPickerProps = SpectrumPickerProps<NormalizedItem>;
 
-export type PickerPropsT<TProps, TChange = ItemKey> = {
-  children: ItemOrSection | ItemOrSection[];
-
-  /** Can be set to true or a TooltipOptions to enable item tooltips */
-  tooltip?: boolean | TooltipOptions;
-
-  /** The currently selected key in the collection (controlled). */
-  selectedKey?: ItemKey | null;
-
-  /** The initial selected key in the collection (uncontrolled). */
-  defaultSelectedKey?: ItemKey;
-
-  /**
-   * Handler that is called when the selection change.
-   * Note that under the hood, this is just an alias for Spectrum's
-   * `onSelectionChange`. We are renaming for better consistency with other
-   * components.
-   */
-  onChange?: (key: TChange) => void;
-
-  onOpenChange?: (isOpen: boolean) => void;
-
-  /** Handler that is called when the picker is scrolled. */
-  onScroll?: (event: Event) => void;
-
-  /**
-   * Handler that is called when the selection changes.
-   * @deprecated Use `onChange` instead
-   */
-  onSelectionChange?: (key: TChange) => void;
-} /*
- * Support remaining SpectrumPickerProps.
- * Note that `selectedKey`, `defaultSelectedKey`, and `onSelectionChange` are
- * re-defined above to account for boolean types which aren't included in the
- * React `Key` type, but are actually supported by the Spectrum Picker component.
- */ & Omit<
-  TProps,
-  // These props are all re-defined above
-  | 'children'
-  | 'onSelectionChange'
-  | 'selectedKey'
-  | 'defaultSelectedKey'
-  // Excluding `defaultItems` and `items` since we are not currently supporting
-  // a render function as `children`. This simplifies the API for determining
-  // initial scroll position and wrapping items with tooltips.
-  | 'defaultItems'
-  | 'items'
->;
-
 export type TooltipOptions = { placement: PopperOptions['placement'] };
 
 /**
